requirements.txt
selenium
pytest
pytest-html
pytest-xdist
faker
pyyaml